+++ packages/expo-modules-core/ios/JSI/EXJavaScriptObject.mm
@@ -76,7 +76,7 @@
   jsi::Runtime *runtime = [_runtime get];
   jsi::Object *jsThis = _jsObjectPtr.get();
 
-  expo::common::definePropertyOnJSIObject(*runtime, jsThis, [name UTF8String], std::move(*[descriptor get]));
+  expo::common::defineProperty(*runtime, jsThis, [name UTF8String], std::move(*[descriptor get]));
 }
 
 - (void)defineProperty:(nonnull NSString *)name value:(nullable id)value options:(EXJavaScriptObjectPropertyDescriptor)options
@@ -87,7 +87,7 @@
   jsi::Object descriptor = [self preparePropertyDescriptorWithOptions:options];
   descriptor.setProperty(*runtime, "value", expo::convertObjCObjectToJSIValue(*runtime, value));
 
-  expo::common::definePropertyOnJSIObject(*runtime, jsThis, [name UTF8String], std::move(descriptor));
+  expo::common::defineProperty(*runtime, jsThis, [name UTF8String], std::move(descriptor));
 }
 
 #pragma mark - WeakObject